refactor(frontend) : InitPage -> MainPage 변경 및 폴더 정리
