+++ frontend/src/component/ChatInit.js
@@ -1,6 +1,6 @@
 import React from "react";
-import classes from "./ChatInit.module.css";
-const ChatInit = (props) => {
+import classes from "../css/ChatInit.module.css";
+const ChatInit = () => {
   return (
     <div className={classes.wrapper}>
       <div className={classes.ulWrapper}></div>
