--- frontend/src/component/ContentsInit.module.css
@@ -1,14 +1,0 @@
-.wrapper {
-  width: 100%;
-  height: auto;
-  display: flex;
-}
-
-.primary {
-  width: 70%;
-  flex-direction: column;
-}
-
-.secondary {
-  width: 30%;
-}
