--- frontend/src/pages/HelloPage.js
@@ -1,6 +1,0 @@
-const HelloPage = () => {
-  console.log("shit");
-  return <div>shit</div>;
-};
-
-export default HelloPage;
