--- frontend/src/pages/InitPage.js
@@ -1,7 +1,0 @@
-import ContentsInit from "../component/ContentsInit";
-
-const InitPage = (props) => {
-  return <ContentsInit is={props.is}></ContentsInit>;
-};
-
-export default InitPage;
